lambdas/process-journey-step/src/test/java/uk/gov/di/ipv/core/processjourneystep/statemachine/StateMachineInitializerTest.java
package uk.gov.di.ipv.core.processjourneystep.statemachine;

import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.extension.ExtendWith;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.EnumSource;
import uk.gov.di.ipv.core.library.domain.IpvJourneyTypes;
import uk.org.webcompere.systemstubs.environment.EnvironmentVariables;
import uk.org.webcompere.systemstubs.jupiter.SystemStub;
import uk.org.webcompere.systemstubs.jupiter.SystemStubsExtension;

import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;

@ExtendWith(SystemStubsExtension.class)
class StateMachineInitializerTest {

    @SystemStub private static EnvironmentVariables environmentVariables;

    @BeforeAll
    private static void beforeAll() {
        environmentVariables.set("IS_LOCAL", "true");
    }

    @ParameterizedTest
    @EnumSource
    void stateMachineInitializerShouldHandleAllStateFiles(IpvJourneyTypes journeyType) {
        assertDoesNotThrow(() -> new StateMachineInitializer(journeyType).initialize());
    }
}
